move block bindiung processor to hooks/
