+++ packages/block-editor/src/components/rich-text/index.js
@@ -47,7 +47,7 @@ import { getAllowedFormats } from './utils';
 import { Content } from './content';
 import { withDeprecations } from './with-deprecations';
 import { unlock } from '../../lock-unlock';
-import { isItPossibleToBindBlock } from '../block-binding-support';
+import { isItPossibleToBindBlock } from '../../hooks/block-binding-support';
 
 export const keyboardShortcutContext = createContext();
 export const inputEventContext = createContext();
