+++ packages/block-editor/src/hooks/block-binding-support/index.js
