+++ packages/block-editor/src/hooks/block-binding-support/with-block-binding-support.js
